+++ web/admin/src/components/organisms/index.ts
@@ -6,3 +6,6 @@ export * from './OrgSidebar';
 export * from './RuntimeStatusPanel';
 export * from './SitesTable';
 export * from './ValidationReportPanel';
+export * from './SiteHeader';
+export * from './SiteTabs';
+export * from './DeploymentUploadPanel';